fix: Expression variable behaviour (PT-184616625)
PT Story: https://www.pivotaltracker.com/story/show/184616625

BUG: When the Edit Variable dialog is open, but none of its fields are in focus, hitting the keyboard Delete key results in deleting the selected variable card.

This change prevents the bug by disabling the default Delete key behavior when a dialog is present.

+++ src/diagram/components/diagram.tsx
@@ -30,6 +30,7 @@ const edgeTypes = {
 };
 
 export interface IProps {
+  dialogPresent?: boolean;
   dqRoot: DQRootType;
   hideControls?: boolean;
   hideNavigator?: boolean;
@@ -42,8 +43,9 @@ export interface IProps {
   showUnusedVariableDialog?: () => void;
   getDiagramExport?: () => unknown;
 }
-export const _Diagram = ({ dqRoot, getDiagramExport, hideControls, hideNavigator, hideNewVariableButton, interactionLocked,
-  setDiagramHelper, showDeleteCardButton, showEditVariableDialog, showUnusedVariableDialog }: IProps) => 
+export const _Diagram = ({ dialogPresent, dqRoot, getDiagramExport, hideControls, hideNavigator,
+  hideNewVariableButton, interactionLocked, setDiagramHelper, showDeleteCardButton,
+  showEditVariableDialog, showUnusedVariableDialog }: IProps) => 
 {
   const reactFlowWrapper = useRef<any>(null);
   const [rfInstance, setRfInstance] = useState<any>();
@@ -106,14 +108,16 @@ export const _Diagram = ({ dqRoot, getDiagramExport, hideControls, hideNavigator
   // onElementsRemove is called when the user's keyboard Delete key is pressed
   // to delete a selected card or a selected connecting arrow (aka "edge").
   const onElementsRemove = (elementsToRemove: Elements) => {
-    for (const element of elementsToRemove) {
-      if (isEdge((element as any))) {
-        const edge = element as Edge;
-        deleteEdge(edge);
-      } else {
-        const nodeToRemove = dqRoot.getNodeFromVariableId(element.id);
-        deleteAllEdgesOfNode(element.id);
-        dqRoot.removeNode(nodeToRemove);
+    if (!dialogPresent) {
+      for (const element of elementsToRemove) {
+        if (isEdge((element as any))) {
+          const edge = element as Edge;
+          deleteEdge(edge);
+        } else {
+          const nodeToRemove = dqRoot.getNodeFromVariableId(element.id);
+          deleteAllEdgesOfNode(element.id);
+          dqRoot.removeNode(nodeToRemove);
+        }
       }
     }
   };
